REinstall react-dom, route code update by jjh
