--- package.json
@@ -1,6 +1,0 @@
-{
-  "dependencies": {
-    "react-router-dom": "^6.3.0",
-    "react-webcam": "^7.0.1"
-  }
-}
