+++ sonsu_web/package.json
@@ -11,6 +11,7 @@
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
     "react-player": "^2.10.1",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
     "web-vitals": "^2.1.4"
   },
