+++ sonsu_web/src/App.js
@@ -1,12 +1,18 @@
 import './App.css';
+import React from "react";
+import { BrowserRouter, Route, Routes } from "react-router-dom";
+import TestMain from './component/TestMain';
 import Test from './component/Test';
-import TestPopup from './component/TestPopup';
 
 function App() {
   return (
     <div className="App">
-        <Test/>
-        <TestPopup/>
+      <BrowserRouter>
+        <Routes path="/" element={<TestMain />}>
+          <Route exact path="test" element={<Test />} />
+        </Routes>
+      </BrowserRouter>
+
     </div>
   );
 }
